requirements.txt
fastapi
uvicorn
ultralytics
pillow
python-multipart
numpy
torch
